feat(log): implement Qodo robustness improvements
Implemented all actionable Qodo code review suggestions:

Logger Improvements:
- Added __attribute__((format(printf, 1, 2))) to LogMessage for compile-time format checking
- Refactored disabled log macros to use LOG_NOOP helper for cleaner code
- Fixed format string bugs caught by new printf checking (2 bugs found!)

SD Card Error Handling:
- Implemented retry loop (max 3 attempts) for partial writes
- Added taskYIELD() between retries to prevent tight loops
- Enhanced error logging with attempt count and bytes written

WiFi Error Handling:
- Added null-checks for pTcpServerContext before dereferencing
- Enhanced bind failure handling with shutdown() and socket invalidation
- Improved error logging with detailed status codes
- Moved UDP_BUFFER_SIZE to file scope per suggestion

Format String Fixes:
- UsbCdc.c: Changed %lu to %u with cast for dropCount
- wifi_tcp_server.c: Changed %lu to %u with cast for errorCount

The printf format attribute immediately caught 2 latent bugs that could
have caused issues at runtime.

+++ firmware/src/services/UsbCdc/UsbCdc.c
@@ -589,8 +589,8 @@ size_t UsbCdc_WriteToBuffer(UsbCdcData_t* client, const char* data, size_t len) 
         // Only log if buffer is critically full to avoid spam
         static uint32_t dropCount = 0;
         if (currentFree < 128 && (++dropCount % 1000) == 0) {
-            LOG_E("USB: Buffer full - dropped %lu packets (needed %d bytes, only %d free)",
-                  dropCount, len, currentFree);
+            LOG_E("USB: Buffer full - dropped %u packets (needed %d bytes, only %d free)",
+                  (unsigned)dropCount, len, currentFree);
         }
         return 0;  // No space available - return immediately
     }
